requirements.txt
pandas>=2.2
numpy>=1.26
yfinance>=0.2.40
statsmodels>=0.14
scipy>=1.11
matplotlib>=3.8
openpyxl>=3.1
python-dotenv>=1.0
requests>=2.31
pytest>=7.4
